refactor: move queries and mutation to its own hooks
