+++ src/components/Workout/WorkoutStarted.tsx
@@ -2,10 +2,6 @@ import React, { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import { v4 as uuidv4 } from 'uuid';
 import { Formik } from 'formik';
-import { useMutation } from '@tanstack/react-query';
-import { DBTable } from '@constants';
-import { useUserContext } from '@contexts/UserContext';
-import { supabase } from '@utils/supabaseClient';
 import { PlusCircleIcon } from '@icons';
 import Button from '@components/Button';
 import WorkoutExerciseCard from '@components/Workout/WorkoutExerciseCard';
@@ -14,6 +10,7 @@ import { Section, WorkoutHeading } from './styles';
 import { validationSchema } from './validationSchema';
 import Timer from './Timer';
 import { useWorkoutContext } from '@contexts/WorkoutContext';
+import { useMutationAddWorkout } from '@hooks/useMutationAddWorkout';
 
 const StyledButton = styled(Button)`
   display: flex;
@@ -64,7 +61,6 @@ type Props = {
 };
 
 const WorkoutStarted: React.FC<Props> = ({ workoutName, setWorkoutName }) => {
-  const { user } = useUserContext();
   const {
     workoutStorage,
     setWorkoutExercises,
@@ -78,33 +74,7 @@ const WorkoutStarted: React.FC<Props> = ({ workoutName, setWorkoutName }) => {
   // Start time in seconds
   const startTime = workoutStorage?.startTime || Date.now() / 1000;
 
-  const mutation = useMutation(
-    async () => {
-      if (!user) {
-        throw new Error('No user found');
-      }
-
-      // Total workout time in seconds
-      const workoutTime = Math.floor(Date.now() / 1000 - startTime);
-      const { data, error } = await supabase.from(DBTable.WORKOUTS).insert({
-        exercises: JSON.stringify(exercises),
-        workoutName,
-        isTemplate: false,
-        workoutTime,
-        userId: user.id,
-      });
-
-      if (error) {
-        throw new Error(error.message);
-      }
-      return data;
-    },
-    {
-      onSuccess: () => {
-        clearWorkoutStorage();
-      },
-    },
-  );
+  const mutation = useMutationAddWorkout({ startTime, exercises });
 
   useEffect(() => {
     setWorkoutExercises(exercises);
